+++ src/components/StoreCard.vue
@@ -1,64 +1,61 @@
 <template>
   <div class="store-card-outer">
 
-    <div v-for="(storeCard, index) in storeCards" :key="index">
+    <div class="picture-outer">
+      <div class="thumbnail" style="background-image:url('https://images.unsplash.com/photo-1506260408121-e353d10b87c7?ixlib=rb-1.2.1&q=85&fm=jpg&crop=entropy&cs=srgb&w=1600&fit=max&ixid=eyJhcHBfaWQiOjE0NTg5fQ')"></div>
+      <div class="thumbnail" style="background-image:url('https://images.unsplash.com/photo-1523712999610-f77fbcfc3843?ixlib=rb-1.2.1&q=85&fm=jpg&crop=entropy&cs=srgb&w=1600&fit=max&ixid=eyJhcHBfaWQiOjE0NTg5fQ')"></div>
+      <div class="thumbnail" style="background-image:url('https://images.unsplash.com/photo-1524260855046-f743b3cdad07?ixlib=rb-1.2.1&q=85&fm=jpg&crop=entropy&cs=srgb&w=1600&fit=max&ixid=eyJhcHBfaWQiOjE0NTg5fQ')"></div>
+      <div class="thumbnail" style="background-image:url('https://images.unsplash.com/photo-1526080676457-4544bf0ebba9?ixlib=rb-1.2.1&q=85&fm=jpg&crop=entropy&cs=srgb&w=1600&fit=max&ixid=eyJhcHBfaWQiOjE0NTg5fQ')"></div>
+    </div>
 
-      <div class="picture-outer">
-        <div>pic1</div>
-        <div>pic2</div>
-        <div>pic3</div>
-        <div>pic4</div>
+    <div class="information-outer">
+      <div class="left-part">
+        <div class="line1">
+          <span class="title">{{ storeCard.storeNameKor }} </span> 
+          <span class="title">{{ storeCard.storeBranchKor }} </span> 
+          <span>{{ storeCard.category }}</span>
+        </div>
+        <div class="line2">
+          <span>{{ storeCard.storeNameEng }} </span>
+          <span>{{ storeCard.storeBranchEng }}</span>
+        </div>
+        <div class="line3">
+          <span class="openNow" v-if="storeCard.openClosed">
+          Open Now</span>
+          <span class="closedNow" v-else>
+          Closed Now</span>
+          <span> · 영업 종료 </span>
+          <span>{{ storeCard.operatingHour }}</span>
+        </div>
+        <div class="line4">
+          <span><img src="@/assets/icon/Icon_seats_cnt.svg"/></span>
+          <span> {{ storeCard.seatsCnt }} · </span>
+          <span><img src="@/assets/icon/Icon_review_cnt.svg"/></span>
+          <span> {{ storeCard.reviewCnt }} · </span>
+          <span><img src="@/assets/icon/Icon_heart_cnt.svg"/></span>
+          <span> {{ storeCard.pinCnt }}</span>
+        </div>
       </div>
 
-      <div class="information-outer">
-        <div class="left-part">
-          <div class="line1">
-            <span class="title">{{ storeCard.storeNameKor }} </span> 
-            <span class="title">{{ storeCard.storeBranchKor }} </span> 
-            <span>{{ storeCard.category }}</span>
-          </div>
-          <div class="line2">
-            <span>{{ storeCard.storeNameEng }} </span>
-            <span>{{ storeCard.storeBranchEng }}</span>
-          </div>
-          <div class="line3">
-            <span class="openNow" v-if="storeCard.openClosed">
-            Open Now</span>
-            <span class="closedNow" v-else>
-            Closed Now</span>
-            <span> · 영업 종료 ▶ </span>
-            <span>{{ storeCard.operatingHour }}</span>
-          </div>
-          <div class="line4">
-            <span><img src="@/assets/icon/Icon_seats_cnt.svg"/></span>
-            <span> {{ storeCard.seatsCnt }} · </span>
-            <span><img src="@/assets/icon/Icon_review_cnt.svg"/></span>
-            <span> {{ storeCard.reviewCnt }} · </span>
-            <span><img src="@/assets/icon/Icon_heart_cnt.svg"/></span>
-            <span> {{ storeCard.pinCnt }}</span>
-          </div>
+      <div class="right-part">
+        <div class="emoji">
+          <div class="emo">🤔<div>10</div></div>
+          <div class="emo">😐<div>200</div></div>
+          <div class="emo">😊<div>888</div></div>
+          <div class="emo">😍<div>1209</div></div>
         </div>
-
-        <div class="right-part">
-          <div class="emoji">
-            <div class="emo"><img src="@/assets/icon/Icon_heart_cnt.svg"/><div>10</div></div>
-            <div class="emo"><img src="@/assets/icon/Icon_heart_cnt.svg"/><div>200</div></div>
-            <div class="emo"><img src="@/assets/icon/Icon_heart_cnt.svg"/><div>888</div></div>
-            <div class="emo"><img src="@/assets/icon/Icon_heart_cnt.svg"/><div>1209</div></div>
-          </div>
-          <div class="distance_pin">
-            <span>{{ storeCard.distance }} km</span>
-            <span @click="pinChange">
-              {{ storeCard.heart }}
-              <span v-if="storeCard.heart">
-                <img src="@/assets/icon/Icon_heart_fill_pin.svg"/></span>
-              <span v-else>
-                <img src="@/assets/icon/Icon_heart_empty_pin.svg"/></span>
-              </span>
+        <div class="distance-pin">
+          <div class="distance">{{ storeCard.distance }} km</div>
+          <div class="pin" @click="pinChange">
+            <div class="pin-img" v-if="storeCard.heart">
+              <img src="@/assets/icon/Icon_heart_fill_pin.svg"/></div>
+            <div class="pin-img" v-else>
+              <img src="@/assets/icon/Icon_heart_empty_pin.svg"/></div>
           </div>
         </div>
       </div>
     </div>
+    <div class="horizontal-line"></div>
 
   </div>
 </template>
@@ -66,18 +63,12 @@
 <script>
   export default {
     props: {
-      storeCards:{
-        type: Array
-      },
-      pin: {
-        type: Boolean,
-        default: false,
+      storeCard:{
       },
     },
-    
     methods: {
       pinChange() {
-        this.$emit('heart', !storeCard.heart)
+        this.$emit('pinChange')
       }
     }
   }
@@ -85,28 +76,32 @@
 
 <style lang="scss" scoped>
   .store-card-outer {
-    border: 1px solid black;
     font-size: 12px;
+
+    .horizontal-line {
+      height: 10px;
+      background-color: #EBEBEB;
+    }
+
     .picture-outer {
-      border: 1px solid green;
       display: flex;
       flex-direction: row;
       justify-content: space-between;
-      padding: 15px 15px 10px 15px;
-      width: calc(100%-30px);
+      padding: 15px 5vw 15px 5vw;
+      width: 90vw;
       :nth-child(n) {
-        border: 1px solid black;
         width: 20vw;
         height: 20vw;
+        background-size: cover;
+        background-repeat: no-repeat;
+        background-position: center center;
       }
     }
     .information-outer {
       display: flex;
-      border: 1px solid orange;
-      padding: 0px 15px 15px 15px;
+      padding: 0px 5vw 15px 5vw;
       .left-part {
-        border: 1px solid red;
-        width: 65vw;
+        width: 60vw;
         .title {
           font-size: 15px;
         }
@@ -131,30 +126,31 @@
         display: flex;
         flex-direction: column;
         justify-content: space-between;
-        border: 1px solid blue;
-        width: 35vw;
+        width: 30vw;
         .emoji {
           display: flex;
           justify-content: space-around;
-          border: 3px solid black;
           .emo {
-            border: 1px solid black;
             font-size: 10px;
             text-align: center;
             width: 25px;
-            img {
-              margin-left: auto;
-              margin-right: auto;       
-            }
             .emoji-temp {
               width: 15px;
               height: 15px;
             }
           }
-
          }
-        .distance_pin {
-          border: 3px solid purple;
+        .distance-pin {
+          display: flex;
+          justify-content: flex-end;
+          .distance {
+            padding: 14px 10px 0px 0px;
+          }
+          .pin {
+            .pin-img {
+              display: flex;
+            }
+          }
         }
       }
     }
